[INDY-1860] move init to separate helper class
